edits

+++ pipeline/sort_files.py
@@ -10,7 +10,7 @@ import pandas as pd
 # Get the user login name
 user_name = os.getlogin()
 
-spcacecraft_data_folder = f"/home/{user_name}/Desktop/git/Lexi-BU/lxi_gui/data/from_LEXI/orbit/"
+spcacecraft_data_folder = f"/home/{user_name}/Desktop/git/Lexi-BU/lxi_gui/data/from_LEXI/surface/"
 spcacecraft_data_folder = Path(spcacecraft_data_folder).expanduser().resolve()
 level_zero_folder = f"/home/{user_name}/Desktop/git/Lexi-BU/lexi_data_pipeline/data/level_0/"
 level_zero_folder = Path(level_zero_folder).expanduser().resolve()
@@ -31,5 +31,12 @@ for file_val in file_val_list[0:]:
     folder_path.mkdir(parents=True, exist_ok=True)
 
     # Copy the file to the new folder
-    shutil.copy(file_val, folder_path / file_name)
-    print(f"Copied {file_name} to {folder_path}")
+    # If the file already exists, skip it
+    file_name = os.path.basename(file_val)
+    if (folder_path / file_name).exists():
+        # print(f"File {file_name} already exists in {folder_path}. Skipping.")
+        continue
+    else:
+        # Copy the file to the new folder
+        shutil.copy(file_val, folder_path / file_name)
+        print(f"Copied {file_name} to {folder_path}")
